refactor: share cephalon managers
