+++ services/ts/cephalon/src/bot.ts
@@ -11,7 +11,7 @@ import {
 import EventEmitter from 'events';
 import { AIAgent } from './agent/index.js';
 import { AGENT_NAME, DESKTOP_CAPTURE_CHANNEL_ID } from '../../../../shared/js/env.js';
-import { ContextManager } from './contextManager';
+import { ContextManager } from '@shared/ts/contextManager';
 import { LLMService } from './llm-service';
 import { interaction, type Interaction } from './interactions';
 import {
